Major Refactor of fetch.py
- Break up fetch.py into logical modules
- Create a directory with `api` related functions
- Create a directory with `processing` related functions
